+++ frontend/package.json
@@ -27,7 +27,7 @@
     "@smui/list": "7.0.0-beta.0",
     "@smui/paper": "7.0.0-beta.0",
     "@smui/ripple": "7.0.0-beta.0",
-    "@smui/segmented-button": "7.0.0-beta.0",
+    "@smui/segmented-button": "^7.0.0-beta.0",
     "@smui/select": "7.0.0-beta.0",
     "@smui/textfield": "7.0.0-beta.0",
     "@smui/tooltip": "7.0.0-beta.0",
