+++ frontend/src/general/Header.svelte
@@ -8,7 +8,6 @@
 		mdiCircleOutline,
 		mdiApi,
 		mdiGithub,
-		mdiDotsCircle,
 	} from "@mdi/js";
 
 	import { updateTab } from "../util/util";
@@ -42,18 +41,6 @@
 			src="build/zeno.png"
 			alt="Square spiral logo next to 'Zeno'" />
 		<div id="tabs">
-			<div
-				class="item {$tab === 'discover' ? 'selected' : ''}"
-				on:click={() => updateTab("discover")}>
-				<div class="icon">
-					<Icon component={Svg} viewBox="0 0 24 24">
-						<path
-							fill={$tab === "discover" ? "#6a1b9a" : "black"}
-							d={mdiDotsCircle} />
-					</Icon>
-				</div>
-				<div class="tab-text">Discover</div>
-			</div>
 			<div
 				class="item {$tab === 'explore' ? 'selected' : ''}"
 				on:click={() => updateTab("explore")}>
