+++ frontend/src/instance-views/scatter/AutoScaleRegl.svelte
