+++ frontend/src/instance-views/scatter/Scatter.svelte
