+++ frontend/src/instance-views/scatter/scatter.ts
